+++ Reviews.h
@@ -1,3 +1,4 @@
+#pragma once
 #include <string>
 using namespace std;
