+++ main.cpp
@@ -1,6 +1,13 @@
-#include "User.h"
-#include "Book.h"
+#include "AccessManager.h"
+#include <iostream>
 
 using namespace std;
+
+int main(){
+    AcessManager control;
+    control.SignUp();
+    control.LoginIn();
+    return 0;
+}
